% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement.R
\name{agreement}
\alias{agreement}
\alias{plot.gosset_agree}
\title{Agreement between ranks}
\usage{
agreement(baseline, compare, labels = NULL)

\method{plot}{gosset_agree}(x, ...)
}
\arguments{
\item{baseline}{a rankings object of class 'rankings' or 'grouped_rankings' that will
be the baseline for comparing the other characteristics}

\item{compare}{a list of objects of same class and dimensions of \code{baseline} to be compared}

\item{labels}{a character to specify the name of compared chacteristics}

\item{x}{object of class 'gosset_agree' for the plotting method. Generates a 'ggplot' object that can be passed to any ggplot2 method}

\item{...}{further arguments passed to methods. Not enabled yet}
}
\value{
A data.frame with summary of agreement:
\item{labels}{the labels for each characteristic}
\item{kendall}{relative Kendall rank correlation coefficient}
\item{first}{relative agreement of a certain item being ranked first in the baseline}
\item{last}{relative agreement of a certain item being ranked last in the baseline}
}
\description{
Summarise the concordance between two or more caracteristics from a baseline
ranking
}
\examples{
# from the breadwheat data
# Compare the overall performance against 
# rankings on germination, grain quality and yield

data("breadwheat", package = "gosset")

R <- rank_tricot(data = breadwheat,
                 items = c(1:3),
                 input = c(18:19))


compare <- list()

compare[[1]] <- rank_tricot(data = breadwheat,
                            items = c("variety_a", "variety_b", "variety_c"),
                            input = c("germination_best","germination_worst"))

compare[[2]] <- rank_tricot(data = breadwheat,
                            items = c("variety_a", "variety_b", "variety_c"),
                            input = c("grainquality_best", "grainquality_worst"))


compare[[3]] <- rank_tricot(data = breadwheat,
                            items = c("variety_a", "variety_b", "variety_c"),
                            input = c("yield_best", "yield_worst"))


labels <- c("Germination", "Grain quality", "Yield")


agreement(R, 
          compare = compare, 
          labels = labels)
 
}
\seealso{
\code{\link{kendallTau}}
}
