% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{AIC.bttree}
\alias{AIC.bttree}
\title{AIC from a Bradley-Terry model
code adapted from PlackettLuce repository
Turner et al (2020)}
\usage{
\method{AIC}{bttree}(object, newdata = NULL, ...)
}
\description{
AIC from a Bradley-Terry model
code adapted from PlackettLuce repository
Turner et al (2020)
}
