% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/akaike_weights.R
\name{akaike_weights}
\alias{akaike_weights}
\title{Akaike weights}
\usage{
akaike_weights(object)
}
\arguments{
\item{object}{a numerical vector with models goodness of fit coefficients}
}
\value{
A data frame containing the coefficients:
\item{delta}{the delta overall change in the coefficients}
\item{relative_logLik}{the relative log-likelihood}
\item{akaike_weights}{the Akaike weights}
}
\description{
Akaike weights represent the relative likelihood of a model. It can be used in model averaging and selection.
}
\examples{

data("airquality")

# try three model approaches
mod1 <- glm(Temp ~ 1,
            data = airquality,
            family = poisson())

mod2 <- glm(Temp ~ Ozone,
            data = airquality,
            family = poisson())

mod3 <- glm(Temp ~ Ozone + Solar.R,
            data = airquality,
            family = poisson())


akaike_weights(c(AIC(mod1), AIC(mod2), AIC(mod3)))
 
}
