% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{bagging}
\alias{bagging}
\title{Bagging}
\usage{
bagging(formula, data, K = 100, N = 1000, seed = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be 
coerced to that class): a symbolic description of the model to be fitted,
of the form \eqn{y ~ x1 + ... + xn}}

\item{data}{a data frame (or object coercible by as.data.frame to a data frame)
containing the variables in the model}

\item{K}{an integer for the number of models in the bagging
\pkg{gosset} will set the seed randomly}

\item{N}{an integer for the number of observations in each bag}

\item{...}{additional arguments passed the methods of the chosen model}
}
\value{
an object of class \code{gosset_cv} with the cross-validation
}
\description{
Methods for bagging
}
\seealso{
Other model selection functions: 
\code{\link{btpermute}()},
\code{\link{crossvalidation}()},
\code{\link{forward}()}
}
\author{
Kauê de Sousa, Jacob van Etten and David Brown
}
\concept{model selection functions}
