% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regret.R
\name{bayes_boot}
\alias{bayes_boot}
\title{Performs a Bayesian bootstrap}
\usage{
bayes_boot(
  data,
  statistic,
  n1 = 1000,
  n2 = 1000,
  use_weights = FALSE,
  weight_arg = NULL,
  ...
)
}
\arguments{
\item{data}{The data as either a vector, matrix or data.frame}

\item{statistic}{A function that accepts data as its first argument and possibly, 
the weights as its second, if use_weights is \code{TRUE}. Should return a numeric vector}

\item{n1}{The size of the bootstrap sample}

\item{n2}{The sample size used to calculate the statistic each bootstrap draw}

\item{use_weights}{Whether the statistic function accepts a weight argument or
should be calculated using resampled data}

\item{weight_arg}{If the statistic function includes a named argument for the
weights this could be specified here}

\item{...}{Further arguments passed on to the statistic method}
}
\description{
Function obtained from Rasmus B책책th's blog at 
https://www.sumsar.net/blog/2015/07/easy-bayesian-bootstrap-in-r/
}
\examples{
# Case 1
set.seed(1337)
exp_data <- rexp(8, rate = 1)
exp_data
bb_sample <- bayes_boot(exp_data, mean, n1 = 10000, n2 = 1000)

# Case 2
boot_fn <- function(cars, weights) {
  loess(dist ~ speed, cars, weights = weights)$fitted
}

bb_loess <- bayes_boot(cars, boot_fn, n1 = 1000, use_weights = TRUE, weight_arg = "weights")

# Plotting the data
plot(cars$speed, cars$dist, pch = 20, col = "tomato4", xlab = "Car speed in mph",
     ylab = "Stopping distance in ft", main = "Speed and Stopping distances of Cars")

# Plotting a scatter of Bootstrapped LOESS lines to represent the uncertainty.
for(i in sample(nrow(bb_loess), 20)) {
  lines(cars$speed, bb_loess[i,], col = "gray")
}
# Finally plotting the posterior mean LOESS line
lines(cars$speed, colMeans(bb_loess, na.rm = TRUE), type ="l",
      col = "tomato", lwd = 4)
 
 @noRd
}
\author{
Rasmus B책책th
}
