% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadwheat.R
\docType{data}
\name{breadwheat}
\alias{breadwheat}
\title{Preferred bread wheat varieties in India}
\format{A data frame with 493 records and 8 variables:
\describe{
    \item{\code{variety_a}}{The name of variety A in the comparison.}
    \item{\code{variety_b}}{The name of variety B in the comparison.}
    \item{\code{variety_c}}{The name of variety C in the comparison.}
    \item{\code{planting_date}}{The sowing date in which the experiment started.}
    \item{\code{lon}}{The longitude in which the experiment was established.}
    \item{\code{lat}}{The latitude in which the experiment was established.}
    \item{\code{best}}{The variety the farmer ranked in first place ("A",
    "B" or "C").}
    \item{\code{worst}}{The variety the farmer ranked in last place ("A",
    "B" or "C").}
}}
\source{
The data were provided by Bioversity International, a CGIAR Research
Centre \url{https://www.bioversityinternational.org}.
}
\usage{
breadwheat
}
\description{
This is a subset of data from crowdsourcing citizen-science trials of
bread wheat (\emph{Triticum aestivum}) varieties in Vaishali, India over the 2014's Rabi.
Farmers were asked to try three varieties of bread wheat randomly assigned
from a total of 16 varieties and evaluate which one had the best and worst 
overall performance. 
  
Environmental data "wheat_rain" and "wheat_temp" are also attached to this 
dataset. These datasets contain the precipitation and temperature information 
extracted from CHIRPS (citation) and MODIS (citation) for a timespan 
of 198 days after the first experiment started.
}
\keyword{datasets}
