% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadwheat_chirps.R
\docType{data}
\name{breadwheat_chirps}
\alias{breadwheat_chirps}
\title{CHIRPS data from preferred bread wheat varieties trial locations}
\format{A matrix with 493 records and 198 colunms:
\describe{
    \item{\code{rows}}{The precipitation data (mm) for each observation in the breadwheat data.}
    \item{\code{colunms}}{Days, from 2014-10-31 to 2015-05-16.}
}}
\source{
The data were obtained from CHIRPS \url{http://chg.geog.ucsb.edu/data/chirps/}.
}
\usage{
breadwheat_chirps
}
\description{
Data from the Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS)
containing the time series precipitation data of the bread wheat 
(\emph{Triticum aestivum}) citizen-science trails during the 2014's Rabi season. The data 
covers a timespan of 198 days from the date that the first experiment started.
}
\seealso{
\code{\link{breadwheat}}
}
\keyword{datasets}
