% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadwheat_modis.R
\docType{data}
\name{breadwheat_modis}
\alias{breadwheat_modis}
\title{MODIS data from preferred bread wheat varieties trial locations}
\format{A array with 493 records, 198 colunms and 2 layers:
\describe{
    \item{\code{rows}}{The temperature data for each observation in the breadwheat data.}
    \item{\code{colunms}}{Days, from 2014-10-31 to 2015-05-16.}
    \item{\code{layer 1}}{The day temperature in Celsius degrees.}
    \item{\code{layer 2}}{The night temperature in Celsius degrees.}
}}
\source{
The data were obtained from MODIS (MYD11A2) \url{http://dx.doi.org/10.5067/MODIS/MYD11A2.006}.
}
\usage{
breadwheat_modis
}
\description{
Data from MODIS (MYD11A2) containing the time series temperature data of the bread wheat 
(\emph{Triticum aestivum}) citizen-science trails during the 2014's Rabi season. The data 
covers a timespan of 198 days from the date that the first experiment started.
}
\seealso{
\code{\link{breadwheat}}
}
\keyword{datasets}
