% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpermute.R
\name{btpermute}
\alias{btpermute}
\title{Variable selection with pair comparisons}
\usage{
btpermute(
  contests = NULL,
  predictors = NULL,
  iterations = 15,
  seed = NULL,
  ...
)
}
\arguments{
\item{contests}{a data frame with pairwise binary contests with these variables
'id','player1','player2','win1','win2'; the id should match with those found in 
\code{predictors}}

\item{predictors}{a data frame with player predictors with an id 
that should match with \code{contests}}

\item{iterations}{integer, number of iterations to compute}

\item{seed}{integer, the seed for random number generation. If NULL (the default), 
\pkg{gosset} will set the seed randomly}

\item{...}{additional arguments passed to \pkg{BradleyTerry2} methods}
}
\value{
an object of class \code{gosset_btpermute} with the final \code{BTm()} model,
selected variables, seeds (random numbers) used and deviances
}
\description{
Permutation method to select variables for pair comparison data
}
\examples{
require("BradleyTerry2")

data("kenyachoice", package = "gosset")

mod <- btpermute(contests = kenyachoice$contests,
                 predictors = kenyachoice$predictors,
                 iterations = 10,
                 seed = 1)

mod
}
\seealso{
\code{\link{rank_binomial}}, \code{\link[BradleyTerry2]{BTm}}

Other model selection functions: 
\code{\link{crossvalidation}()},
\code{\link{forward}()}
}
\concept{model selection functions}
