% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeinfo.R
\name{build_tree_nodes}
\alias{build_tree_nodes}
\title{Build tree nodes
This function makes the panels}
\usage{
build_tree_nodes(
  x,
  log = TRUE,
  ref = NULL,
  ci.level = 0.95,
  node.ids = NULL,
  n.obs = NULL
)
}
\arguments{
\item{x}{a list with PlackettLuce objects}

\item{log}{logical, if \code{TRUE} log-worth coefficients are 
displayed instead of worth}

\item{ref}{optional, character for the reference item when
\var{log} = \code{TRUE}}

\item{ci.level}{an integer for the confidence interval levels}
}
\description{
Build tree nodes
This function makes the panels
}
