% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{crossvalidation}
\alias{crossvalidation}
\title{Cross-validation}
\usage{
crossvalidation(formula, data, k = NULL, folds = NULL,
  mean.method = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted,
of the form \eqn{ y ~ x1 + ... + xn }}

\item{data}{a data frame, list or environment
(or object coercible by as.data.frame to a data frame) 
containing the variables in the model.}

\item{k}{an integer for the number of splits in the cross-validation}

\item{folds}{an optional vector specifying the folds in the cross-validation}

\item{mean.method}{a character for the method to calculate the mean of cross-validation estimators. 
Options are: 'equal', an arithmetic mean; 
'foldsize', a weighted mean by the size in each fold; 
'stouffer' a weighted Z-test developed by Stouffer et al. (1949). 
The two last methods are suggested for cross-validation with imbalaced folds}

\item{...}{additional arguments passed to methods}
}
\value{
The cross-validation goodness-of-fit estimates for the best model, which are:
\item{AIC}{Akaike Information Criterion}
\item{deviance}{Model deviance}
\item{logLik}{Log-Likelihood}
\item{MaxLik}{Maximum likelihood pseudo R-squared}
\item{CraggUhler}{Cragg and Uhler's pseudo R-squared}
\item{Agresti}{Agresti pseudo R-squared}
Cross-validation estimates are computed using the fitted models on the validation samples.
}
\description{
Methods for measuring the performance of a predictive model on sets of test data.
}
\examples{

# Generalized Linear Models

data("airquality")

crossvalidation(Temp ~ Wind + Solar.R,
                data = airquality,
                k = 5,
                family = poisson())

########################################

# PlackettLuce Model
# beans data from PlackettLuce
library("PlackettLuce")

example("beans", package = "PlackettLuce")

G <- grouped_rankings(R, rep(seq_len(nrow(beans)), 4))

beans <- cbind(G, beans)

# take seasons as bins
k <- length(unique(beans$season))
folds <- as.integer(as.factor(beans$season))

crossvalidation(G ~ maxTN, 
                data = beans, 
                k = k, 
                folds = folds, 
                minsize = 100)

########################################

# BradleyTerry Model
library("psychotree")
# Germany's Next Topmodel 2007 data from psychotree
data("Topmodel2007", package = "psychotree")

crossvalidation(preference ~ ., 
                data = Topmodel2007,
                k = 5)
                
}
\seealso{
\code{\link[gnm]{gnm}}, \code{\link[PlackettLuce]{pltree}}, 
\code{\link[psychotree]{bttree}}
}
