% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_tricot.R
\name{.validate_rankings}
\alias{.validate_rankings}
\title{Validate rankings}
\usage{
.validate_rankings(x)
}
\arguments{
\item{x}{data.frame with two columns indicating the tricot rankings}
}
\description{
This check ranking consistency making sure that 
no NAs or ties are mantained in the final PlackettLuce ranking
}
