% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Forward stepwise regression for model selection}
\usage{
forward(formula, data, k = NULL, folds = NULL, select.by = NULL,
  ncores = NULL, packages = NULL, ...)
}
\arguments{
\item{formula}{a symbolic description of the model, if set as
\eqn{ y ~ . }   all variables in data are used}

\item{data}{a data frame, list or environment
(or object coercible by as.data.frame to a data frame) 
containing the variables in the model.}

\item{k}{an integer for the number of splits in the cross-validation}

\item{folds}{an optional vector specifying the folds in the cross-validation}

\item{select.by}{a character for the goodness-of-fit statistical parameter to
select the models. Set as 'deviance' by default.}

\item{ncores}{a integer for the number of cores to be used in the parallel computing. 
Set as 1 by default (no parallelisation)}

\item{packages}{an optional character vector of packages that the parallel tasks depend on, ignore
if all required packages are loaded to .GlobalEnv}

\item{...}{additional arguments passed to the selected fitting model}
}
\value{
The cross-validation goodness-of-fit estimates for the best model, which are:
\item{AIC}{Akaike Information Criterion}
\item{deviance}{Model deviance}
\item{logLik}{Log-Likelihood}
\item{MaxLik}{Maximum likelihood pseudo R-squared}
\item{CraggUhler}{Cragg and Uhler's pseudo R-squared}
\item{Agresti}{Agresti pseudo R-squared}
Cross-validation estimates are computed using the fitted models on the validation samples.
}
\description{
Forward selection is a type of stepwise regression which begins with an empty 
model and adds in variables one by one. In each step one variable is added
when it makes a single best improvement to the model.
}
\examples{

library("gnm")
library("doParallel")
library("abind")

data("airquality")

mod <- forward(Temp ~ .,
               data = airquality,
               k = 5,
               select.by = "AIC",
               family = poisson(link = "log"))


########################################
 
# breadwheat data
library("PlackettLuce")
library("doParallel")
library("abind")

data("breadwheat", package = "gosset")

# data frame into a object of class 'rankings' from PlackettLuce
G <- to_rankings(breadwheat, 
                 items = c(1:3), 
                 rankings = c(7:8),
                 type = "tricot",
                 grouped.rankings = TRUE)

# get some explanatory variables to include in the model
# we will calculate the temperature indices
# for the planting date to maturity (120 days)

data("breadwheat_modis", package = "gosset")

covars <- temperature(breadwheat_modis,
                      day.one = breadwheat$planting_date,
                      span = 120)

# combine the grouped_rankings and the explanatory variables
data <- cbind(G, covars, breadwheat[c("lon","lat")])

# run forward selection
mod <- forward(G ~ ., 
               data = data, 
               k = 3, 
               ncores = 2,
               minsize = 50, 
               alpha = 0.01)
               
}
\seealso{
\code{\link{crossvalidation}}
}
