% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{get_rules_labels}
\alias{get_rules_labels}
\title{Get labels of variables used in a party tree}
\usage{
get_rules_labels(tree, var = NULL)
}
\arguments{
\item{var}{optional, a vector of existing variables used to fit 
a party tree in \var{x}, model.frame() is used by default}

\item{x}{an object of class party}
}
\description{
Get labels of variables used in a party tree
}
\examples{
example("beans", package = "PlackettLuce")
G <- group(R, rep(seq_len(nrow(beans)), 4))

tree <- pltree(G ~ maxTN + season,  data = beans, alpha = 0.05)

get_rules_labels(tree)
}
