% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_cm.R
\name{hide_cm}
\alias{hide_cm}
\title{Hide in-code comments of specified number of hashtags}
\usage{
hide_cm(hash = "###", switch = TRUE)
}
\arguments{
\item{hash}{character of hashes to indicate which comments to hide}

\item{switch}{a Boolean. Switch for on and off.}
}
\description{
Hide in-code comments of specified number of hashtags
}
\examples{
hide_cm(hash = "##")
}
