% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.R
\name{ind}
\alias{ind}
\title{Set Indentation for text}
\usage{
ind(p = 24)
}
\arguments{
\item{p}{a number for points to indent by.}
}
\description{
Points for paragraph indentation
}
\examples{
ind()
}
\author{
Daniel Hintz
}
