% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/utility.R
\name{is_decimal}
\alias{is_decimal}
\title{Decimal number}
\usage{
is_decimal(x)
}
\arguments{
\item{x}{vector with numbers of class "numeric" or "integer"}
}
\value{
a logical vector
}
\description{
Find a decimal number in a vector with integers
}
\examples{

x <- c(1, 2, 1.4, 3, 4.01)

is_decimal(x)

}
