% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/utility.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Whole number}
\usage{
is_wholenumber(x)
}
\arguments{
\item{x}{vector with numbers of class "numeric" or "integer"}
}
\value{
a logical vector
}
\description{
Find integer values in a vector with decimals
}
\examples{

x <- c(1, 2, 1.4, 3, 4.01)

is_wholenumber(x)

}
