% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_permute.R
\name{kendall_permute}
\alias{kendall_permute}
\title{Kendall's tau permutation test}
\usage{
kendall_permute(x, y, split, n.permutations = 500)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame}

\item{y}{a vector, matrix or data frame with compatible dimensions to \code{x}}

\item{split}{a vector indicating the splitting rule for the test}

\item{n.permutations}{an integer, the number of permutations to perform}
}
\value{
A data.frame containing:
\item{observed_diff}{observed absolute differences in Kendall's tau for all group pairs.}
\item{p_values}{p-values from the permutation test for all group pairs.}
}
\description{
Perform a pairwise permutation test to assess statistical differences 
in Kendall's Tau correlation between two or more groups.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set.seed(42)
x = rnorm(100)
y = rnorm(100)
split = rep(c("Group1", "Group2", "Group3"), length.out = 100)
kendall_permute(x, y, split)

data("breadwheat", package = "gosset")

x = rank_tricot(breadwheat, 
                items = paste0("variety_", letters[1:3]),
                input = c("yield_best", "yield_worst"),
                validate.rankings = TRUE)

y = rank_tricot(breadwheat, 
                items = paste0("variety_", letters[1:3]),
                input = c("overall_best", "overall_worst"),
                validate.rankings = TRUE)
                
kendall_permute(x, y, 
                split = rep(c("Group1", "Group2", "Group3"), length.out = nrow(breadwheat)), 
                n.permutations = 100)
                
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[gosset]{kendallTau}}
}
\author{
Kauê de Sousa
}
