% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallTau.R
\name{kendallTau}
\alias{kendallTau}
\title{Kendall rank correlation coefficient}
\usage{
kendallTau(x, y, null.rm = TRUE)
}
\arguments{
\item{x}{a matrix with rankings or model coefficients with same dimensions of \code{y}}

\item{y}{a matrix with rankings or model coefficients with same dimensions of \code{x}}

\item{null.rm}{logical, to remove zeros from matrix \code{x} and \code{y}}
}
\value{
The Kendall correlation coefficient and the Effective N, which 
is the equivalent N needed if all items were compared to all items. 
Can be used for significance testing.
}
\description{
Compute Kendall rank correlation coefficient between two matrices. Is coefficient 
used in statistics to measure the ordinal association between two measured quantities.
A tau test is a non-parametric hypothesis test for statistical dependence based 
on the tau coefficient.
}
\examples{

# check the correlation between observed rankings 
# and the predicted rankings from PlackettLuce

library("PlackettLuce")

R <- matrix(c(1, 2, 4, 3,
              1, 4, 2, 3,
              1, 2, 4, 3,
              1, 2, 4, 3,
              1, 3, 4, 2,
              1, 4, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) <- LETTERS[1:4]

G <- group(as.rankings(R), 1:6)

mod <- pltree(G ~ 1, data = G)

preds <- predict(mod)

k <- kendallTau(R, preds)

}
\references{
Kendall M. G. (1938). Biometrika, 30(1–2), 81–93. https://doi.org/10.1093/biomet/30.1-2.81.
}
\seealso{
\code{\link[stats]{cor}}
}
