% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenyachoice.R
\docType{data}
\name{kenyachoice}
\alias{kenyachoice}
\title{Kenya choices}
\format{A list}
\source{
The data were provided by Bioversity International \url{https://www.bioversityinternational.org/}, 
a CGIAR Research Centre.
}
\usage{
kenyachoice
}
\description{
Explain
}
\keyword{datasets}
