% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_test.R
\name{LR_test}
\alias{LR_test}
\title{Likelihood-ratio test}
\usage{
LR_test(x, split, ...)
}
\arguments{
\item{x}{a object of class rankings or grouped_rankings}

\item{split}{a vector indicating the splitting rule for the test}

\item{...}{additional arguments passed to methods}
}
\description{
Assesses the goodness of fit of two competing
 statistical models
}
\examples{
library("PlackettLuce")
example("beans", package = "PlackettLuce")
G = group(R, rep(seq_len(nrow(beans)), 4))
d = cbind(G, beans)

split = ifelse(d$maxTN < 18.7175, T, F)

LR_test(G, split)
}
\author{
Joost van Heerwaarden and Kauê de Sousa
}
