% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multcompPL.R
\name{multcompPL}
\alias{multcompPL}
\alias{multcompPL.default}
\alias{multcompPL.pltree}
\title{Post-hoc analysis of Plackett-Luce models}
\usage{
multcompPL(mod, terms = NULL, threshold = 0.05, adjust = "none", ...)

\method{multcompPL}{default}(mod, terms = NULL, threshold = 0.05, adjust = "none", ...)

\method{multcompPL}{pltree}(mod, terms = NULL, threshold = 0.05, adjust = "none", ...)
}
\arguments{
\item{mod}{an object of class PlackettLuce}

\item{terms}{a vector with the models terms to be used}

\item{threshold}{a numeric value, second reference argument to compare}

\item{adjust}{p.value correction method, a character string}

\item{...}{additional arguments passed to methods in qvcalc}
}
\description{
Post-hoc analysis of Plackett-Luce models
}
\examples{
\donttest{
require("psychotree")
require("PlackettLuce")
require("multcompView")

data("Topmodel2007", package = "psychotree")

R <- as.grouped_rankings(Topmodel2007$preference)

plt <- pltree(R ~ ., data = Topmodel2007[, -1], minsize = 5, npseudo = 0)

multcompPL(plt, threshold = 0.05, ref = "Anni")
}
}
