% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeinfo.R
\name{node_labels}
\alias{node_labels}
\alias{node_rules}
\alias{top_items}
\title{Get node labels and rules used in a party tree}
\usage{
node_labels(x)

node_rules(x)

top_items(x, top = 5)
}
\arguments{
\item{x}{an object of class \code{party}}

\item{top}{an integer for the number of items to return}
}
\value{
a vector with the node labels, a data.frame with node rules
}
\description{
Get node labels and rules used in a party tree

Get node rules

Get the top items out of a decision tree
}
\examples{
if (require("PlackettLuce")) {
example("beans", package = "PlackettLuce")
G <- group(R, rep(seq_len(nrow(beans)), 4))

tree <- pltree(G ~ maxTN + season + lon, data = beans)

node_labels(tree)

node_rules(tree)

top_items(tree)

}
}
\author{
Kauê de Sousa
}
