% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference.R
\name{preference}
\alias{preference}
\alias{plot.cmb_pref}
\title{Pairwise preference}
\usage{
preference(data = NULL, items = NULL, input = NULL, ...)

\method{plot}{cmb_pref}(x, ...)
}
\arguments{
\item{data}{a data frame with columns specified by items and input values}

\item{items}{a data frame or index of \code{data} for the column(s) containing the item names}

\item{input}{a data frame or index of \code{data} for the column(s) containing the values to be ranked}

\item{...}{additional arguments passed to methods. See details}

\item{x}{an object of class 'cmb_pref' for the plotting method. Generates a 'ggplot' object that can be passed to any ggplot2 method}
}
\value{
a data.frame with pairwise comparisons 
\item{player1}{the first player in the comparison}
\item{player2}{the second player in the comparison}
\item{win}{times player1 wins against player2}
\item{ncontest}{number of contests between player1 and player2}
\item{preference}{number of times player1 is preferred against player2}
}
\description{
Compute pairwise preference matrix between comparisons
}
\examples{

# breadwheat data

data("breadwheat", package = "gosset")

preference(data = breadwheat,
           items = c(1:3),
           input = c(18:19))

# beans data where each observer compares 3 varieties
# randomly distributed

library("PlackettLuce")
data("beans", package = "PlackettLuce")

preference(data = beans,
           items = c(1:3),
           input = c(4:5))

# also include the local item
preference(data = beans,
           items = c(1:3),
           input = c(4:5),
           additional.rank = beans[c(6:8)])
   
}
