% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.df.hlp.R
\name{print.df.hlp}
\alias{print.df.hlp}
\title{Helper functon for print.df}
\usage{
\method{print}{df.hlp}(x, ...)
}
\arguments{
\item{x}{a dataframe}

\item{...}{an ellipse}
}
\value{
a formated dataframe
}
\description{
Helper functon for print.df
}
\author{
Daniel Hintz
}
