% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.df.R
\name{print.df}
\alias{print.df}
\title{Automatically Format printing for dataframes}
\usage{
\method{print}{df}(x, ...)
}
\arguments{
\item{x}{a dataframe}

\item{...}{an ellispse}
}
\value{
a dataframe
}
\description{
Automatically Format printing for dataframes
}
\author{
Daniel Hintz
}
