% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mat.hlp.R
\name{print.mat.hlp}
\alias{print.mat.hlp}
\title{Helper function for print.mat}
\usage{
\method{print}{mat.hlp}(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{an ellipse}
}
\value{
a vector
}
\description{
Helper function for print.mat
}
