% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mat.R
\name{print.mat}
\alias{print.mat}
\title{Automated Print Formatting for matrices}
\usage{
\method{print}{mat}(x, ...)
}
\arguments{
\item{x}{A matrix}

\item{...}{ellispe not set}
}
\value{
A matrix
}
\description{
Automated Print Formatting for matrices
}
\author{
Daniel Hintz
}
