% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMatHlp.R
\name{printMatHlp}
\alias{printMatHlp}
\title{Helper function for print.mat}
\usage{
printMatHlp(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{an ellipse}
}
\description{
Helper function for print.mat
}
\author{
Daniel Hintz
}
