% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printS3.R
\name{printS3}
\alias{printS3}
\title{Automatically Format Printing}
\usage{
printS3(hash = "###")
}
\arguments{
\item{hash}{a character indicating hashes of which length will be hidden}
}
\description{
Automatically Format Printing
}
\author{
Daniel Hintz
}
