% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoR2.R
\name{pseudoR2}
\alias{pseudoR2}
\title{Pseudo R-squared}
\usage{
pseudoR2(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{additional arguments affecting the R-squared produced}
}
\value{
A data frame containing the pseudo R-squared coefficients:
\item{logLik}{log-likelihood}
\item{logLikNull}{Null log-likelihood}
\item{MaxLik}{maximum likelihood pseudo R-squared}
\item{CraggUhler}{Cragg and Uhler's pseudo R-squared}
\item{Agresti}{Agresti pseudo R-squared}
}
\description{
Regression coefficient to evaluate goodness-of-fit in a given model
}
\examples{

data("airquality")

mod <- glm(Temp ~ Wind + Solar.R,
           data = airquality,
           family = poisson())

pseudoR2(mod)

}
