% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_numeric.R
\name{rank_numeric}
\alias{rank_numeric}
\title{Plackett-Luce rankings from numeric values}
\usage{
rank_numeric(
  data = NULL,
  items = NULL,
  input = NULL,
  id = NULL,
  group = FALSE,
  ascending = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame with columns specified by items and input values}

\item{items}{a data frame or index of \code{data} for the column(s) containing the item names}

\item{input}{a data frame or index of \code{data} for the column(s) containing the values to be ranked}

\item{id}{a data frame or index of \code{data} indicating the ids for "long" \code{data}}

\item{group}{logical, if TRUE return an object of class "grouped_rankings"}

\item{ascending}{logical, to compute rankings from lower to higher values}

\item{...}{additional arguments passed to methods}
}
\value{
a PlackettLuce "rankings" object, which is a matrix of dense rankings
}
\description{
Group and convert numeric values into integer ranks
}
\examples{
# A matrix with 10 rankings of 5 items (A, B, C, D, E)
# with numeric values as "rank"
set.seed(123)
df <- cbind(id = rep(1:10, each = 5),
            items = rep(LETTERS[1:5], times = 10),
            input = runif(50, 1, 3))

# return an object of class 'rankings'
R <- rank_numeric(df,
                  items = 2,
                  input = 3,
                  id = 1)

# rankings can be computed in ascending order
R <- rank_numeric(df,
                  items = 2,
                  input = 3,
                  id = 1,
                  ascending = TRUE)


# return an object of class 'grouped_rankings'
R <- rank_numeric(df,
                  items = 2,
                  input = 3,
                  id = 1,
                  grouped.rankings = TRUE)

######################

# Rankings with 5 items randomly assigned

i <- as.data.frame(matrix(NA, nrow = 10, ncol = 5))
names(i) <- paste0("Item",1:5)

r <- as.data.frame(matrix(NA, nrow = 10, ncol = 5))
names(r) <- paste0("Position_Item",1:5)

for(s in 1:10) {
  i[s,] <- sample(LETTERS[1:5])
  r[s,] <- sample(1:5)
}

R <- rank_numeric(items = i,
                  input = r)
 
}
\seealso{
\code{\link[PlackettLuce]{rankings}}
}
