% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/rank_paircomp.R
\name{rank_paircomp}
\alias{rank_paircomp}
\title{Pairwise comparison rankings}
\usage{
rank_paircomp(object)
}
\arguments{
\item{object}{an object of class "rankings" or "grouped_rankings" which is a matrix of dense rankings.}
}
\value{
an object of class "paircomp" which is a matrix of pairwise comparisons
}
\description{
Pairwise comparisons from a ranking object. Ties are not taken into account, then they are added as NA's.
}
\examples{

library("PlackettLuce")
library("psychotools")

# a simple matrix with 4 items
# ties are computed as NA's 
R <- matrix(c(1, 2, 0, 0,
              4, 1, 2, 3,
              2, 4, 3, 1,
              1, 2, 3, 0,
              2, 1, 1, 0,
              1, 0, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) <- c("apple", "banana", "orange", "pear")
R <- PlackettLuce::as.rankings(R)

rank_paircomp(R)

#############################################

# using breadwheat data
data("breadwheat", package = "gosset")

# convert the tricot rankings from breadwheat data
# into a object of class 'rankings' from PlackettLuce
R <- rank_PL(breadwheat,
             items = c("variety_a","variety_b","variety_c"),
             rankings = c("overall_best","overall_worst"),
             type = "tricot")
 

PC <- rank_paircomp(R)


}
\seealso{
\code{\link[PlackettLuce]{rankings}} , \code{\link[psychotools]{paircomp}}
}
