% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap_dominance.R
\name{recap_dominance}
\alias{recap_dominance}
\alias{plot.gosset_dmnc}
\title{Pairwise player dominance}
\usage{
recap_dominance(object)

\method{plot}{gosset_dmnc}(x, ...)
}
\arguments{
\item{object}{a rankings object of class 'rankings', 'grouped_rankings'
or 'paircomp'}

\item{x}{an object of class 'gosset_dmnc' for the plotting method. 
Generates a 'ggplot' object that can be passed to any ggplot2 method}

\item{...}{further arguments passed to methods. Not enabled yet}
}
\value{
A data.frame with summary of dominance from pairwise comparisons: 
\item{player1}{the first player in the comparison}
\item{player2}{the second player in the comparison}
\item{ncontest}{number of contests between player1 and player2}
\item{dominance}{relative dominance of player1 on player2}
}
\description{
Methods to summarise player dominance from pairwise comparisons
}
\examples{
# breadwheat data

data("breadwheat", package = "gosset")
 
R <- rank_tricot(data = breadwheat,
                 items = c("variety_a", "variety_b", "variety_c"),
                 input = c("overall_best", "overall_worst"))

d <- dominance(R)

p <- plot(d)

# beans data where each observer compares 3 varieties
# randomly distributed

library("PlackettLuce")
data("beans", package = "PlackettLuce")

R <- rank_tricot(data = beans,
                 items = c("variety_a", "variety_b", "variety_c"),
                 input = c("best", "worst"),
                 additional.rank = beans[, c("var_a", "var_b", "var_c")])

d <- recap_dominance(R)

 
}
\seealso{
Other recap functions: 
\code{\link{recap_agreement}()},
\code{\link{recap_favourite}()},
\code{\link{recap_victories}()}
}
\concept{recap functions}
