% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap_victories.R
\name{recap_victories}
\alias{recap_victories}
\alias{plot.gosset_vctr}
\title{Pairwise victories}
\usage{
recap_victories(object)

\method{plot}{gosset_vctr}(x, ...)
}
\arguments{
\item{object}{a rankings object of class 'rankings', 'grouped_rankings' 
or 'paircomp'}

\item{x}{an object of class 'gosset_vctr' for the plotting method. 
Generates a 'ggplot' object that can be passed to any ggplot2 method}

\item{...}{further arguments passed to methods. Not enabled yet}
}
\value{
A data.frame with summary of victories from pairwise comparisons: 
\item{player1}{the first player in the comparison}
\item{player2}{the second player in the comparison}
\item{win}{times player1 wins against player2}
\item{ncontest}{number of contests between player1 and player2}
\item{victories}{relative number of times player1 wins against player2}
}
\description{
Summarise victories from pairwise comparisons
}
\examples{
# breadwheat data

data("breadwheat", package = "gosset")
 
R <- rank_tricot(data = breadwheat,
                 items = c("variety_a", "variety_b", "variety_c"),
                 input = c("overall_best", "overall_worst"))


v <- recap_victories(R)

p <- plot(v)

# beans data where each observer compares 3 varieties
# randomly distributed

library("PlackettLuce")
data("beans", package = "PlackettLuce")
 
R <- rank_tricot(data = beans,
                 items = c("variety_a", "variety_b", "variety_c"),
                 input = c("best", "worst"),
                 additional.rank = beans[, c("var_a", "var_b", "var_c")])

v <- recap_victories(R)

 
}
\seealso{
Other recap functions: 
\code{\link{recap_agreement}()},
\code{\link{recap_dominance}()},
\code{\link{recap_favourite}()}
}
\concept{recap functions}
