% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_agreement.R
\name{summarise_agreement}
\alias{summarise_agreement}
\alias{plot.gosset_agree}
\title{Agreement between rankings}
\usage{
summarise_agreement(baseline, compare.to, labels = NULL)

\method{plot}{gosset_agree}(x, ...)
}
\arguments{
\item{baseline}{an object of class 'rankings' or 'grouped_rankings' 
that serves as baseline for comparing the other characteristics}

\item{compare.to}{a list of objects of same class and dimensions of 
\code{baseline} to be compared}

\item{labels}{a character to specify the name of compared chacteristics}

\item{x}{object of class 'gosset_agree' for the plotting method. 
Generates a 'ggplot' object that can be passed to any ggplot2 method}

\item{...}{further arguments passed to methods. Not enabled yet}
}
\value{
A data.frame with summary of agreement:
\item{labels}{the labels for each characteristic}
\item{kendall}{relative Kendall rank correlation coefficient}
\item{first}{relative agreement of the first item in the baseline being 
ranked first in compare.to}
\item{last}{relative agreement of the last item in the baseline being 
ranked last in compare.to}
}
\description{
Summarise the concordance between one or more caracteristics from a baseline
ranking
}
\examples{
# from the breadwheat data
# Compare the overall performance against
# rankings on germination, grain quality and yield

data("breadwheat", package = "gosset")
 
R <- rank_tricot(data = breadwheat,
                 items = c("variety_a", "variety_b", "variety_c"),
                 input = c("overall_best", "overall_worst"))


compare <- list()

compare[[1]] <- rank_tricot(data = breadwheat,
                            items = c("variety_a", "variety_b", "variety_c"),
                            input = c("germination_best","germination_worst"))

compare[[2]] <- rank_tricot(data = breadwheat,
                            items = c("variety_a", "variety_b", "variety_c"),
                            input = c("grainquality_best", "grainquality_worst"))


compare[[3]] <- rank_tricot(data = breadwheat,
                            items = c("variety_a", "variety_b", "variety_c"),
                            input = c("yield_best", "yield_worst"))


labels <- c("Germination", "Grain quality", "Yield")


summarise_agreement(R,
                    compare.to = compare,
                    labels = labels)
 
}
\seealso{
\code{\link{kendallTau}}

Other summarise functions: 
\code{\link{summarise_dominance}()},
\code{\link{summarise_favourite}()},
\code{\link{summarise_victories}()}
}
\concept{summarise functions}
