% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_sp.R
\name{tex_sp}
\alias{tex_sp}
\title{Spacing for pdf Rmarkdowns}
\usage{
tex_sp(spacing = "single")
}
\arguments{
\item{spacing}{The character "single" or "double"}
}
\description{
Single or double spacing for Rmarkdowns outputting to pdf
}
\details{
Intended use is as a workhorse function for tex(), though can be used independently
}
\examples{
tex_sp()
}
\author{
Daniel Hintz
}
