% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_paircomp.R
\name{to_paircomp}
\alias{to_paircomp}
\title{Pairwise comparison rankings}
\usage{
to_paircomp(object)
}
\arguments{
\item{object}{a "rankings" or "grouped_rankings" object which is a matrix of dense rankings. 
The function do not take ties into account, then they are added as NA's}
}
\value{
an object of class "paircomp" which is a matrix of pairwise comparisons
}
\description{
Pairwise comparisons from a ranking object
}
\examples{

library("psychotools")

# a simple matrix with 4 items
# ties are computed as NA's 
R <- matrix(c(1, 2, 0, 0,
              4, 1, 2, 3,
              2, 4, 3, 1,
              1, 2, 3, 0,
              2, 1, 1, 0,
              1, 0, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) <- c("apple", "banana", "orange", "pear")
R <- as.rankings(R)

to_paircomp(R)

#############################################

# using breadwheat data
data("breadwheat", package = "gosset")

# convert the tricot rankings from breadwheat data
# into a object of class 'rankings' from PlackettLuce
R <- to_rankings(breadwheat,
                 items = c(1:3),
                 rankings = c(7:8),
                 type = "tricot")


to_paircomp(R)

}
\seealso{
\code{\link[PlackettLuce]{rankings}} , \code{\link[psychotools]{paircomp}}
}
