% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rankings.R
\name{to_rankings}
\alias{to_rankings}
\title{Plackett-Luce rankings}
\usage{
to_rankings(data = NULL, items = NULL, rankings = NULL,
  type = NULL, ...)
}
\arguments{
\item{data}{a data frame with columns specified by items and rankings. Data can be in both wide and long format.
An id is required if long format.}

\item{items}{a data frame or vector representing the item names}

\item{rankings}{a data frame or vector representing the rankings}

\item{type}{the type of data input. Options are 'tricot' and 'rank'}

\item{...}{further arguments passed to PlackettLuce::rankings}
}
\value{
a "rankings" object, which is a matrix of dense rankings
}
\description{
Create an object of class "rankings" from a dataframe or matrix
}
\examples{

library("PlackettLuce")

# A matrix with 10 rankings of 5 items (A, B, C, D, E)
# with numeric values as "rank"
set.seed(123)
df <- cbind(id = rep(1:10, each = 5),
            items = rep(LETTERS[1:5], times = 10),
            rankings = runif(50, 1, 3))
df

# return a 'rankings' object
to_rankings(df,
            items = 2,
            rankings = 3,
            id = 1)

# return a 'grouped_rankings' object
to_rankings(df,
            items = 2,
            rankings = 3,
            id = 1,
            grouped.rankings = TRUE)

##################################

# Rankings with 5 items randomly assigned

i <- as.data.frame(matrix(NA, nrow = 10, ncol = 5))
names(i) <- paste0("Item",1:5)

r <- as.data.frame(matrix(NA, nrow = 10, ncol = 5))
names(r) <- paste0("Position_Item",1:5)

for(s in 1:10) {
  i[s,] <- sample(LETTERS[1:5])
  r[s,] <- sample(1:5)
}

cbind(i, r) 


to_rankings(items = i,
            rankings = r)

###################################

# breadwheat data, which is an object ordered in the 'tricot' format
# each observer compares 3 varieties randomly distributed from a list of 16 
data("breadwheat", package = "gosset")

to_rankings(breadwheat,
            items = c("variety_a","variety_b","variety_c"),
            rankings = c("best","worst"),
            type = "tricot")

######################

# beans data which is an object ordered in the 'tricot'format
# each observer compares 3 varieties randomly distributed from a list of 11
# and additionally compares these 3 varieties with their local variety
library("PlackettLuce")
data("beans", package = "PlackettLuce")

# comparison with local item is added as an additional rankings, then
# each of the 3 varieties are compared separately with the local item,
# it return a object four times larger (in rows) than the input data
# combining this with covariates from other dataset is easy since 
# the function keeps an internal id
to_rankings(beans,
            items = c(1:3),
            rankings = c(4:5),
            local = c(6:8),
            type = "tricot",
            grouped.rankings = TRUE)

}
\seealso{
\code{\link[PlackettLuce]{rankings}}
}
