% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worst_regret.R
\name{worst_regret}
\alias{worst_regret}
\title{Worst regret in regret-based models from recursive partitioning trees}
\usage{
worst_regret(object, ...)
}
\arguments{
\item{object}{an object of class modelparty}

\item{...}{further arguments passed to methods}
}
\value{
A data frame with probabilities of winning
\item{items}{the item names}
\item{win_probs}{the probabilities of winning}
\item{worst_regret}{the worst regret index}
}
\description{
Compute a worst regret index from recursive partitioning trees. Regret is an important
heuristic in the behavioural sciences. Minimizing worst regret (the loss under the worst
possible outcome) is a criterion that takes a conservative approach to risk analysis in 
diversification strategies.
}
\examples{

library("PlackettLuce")

data("breadwheat", package = "gosset")

# convert the tricot rankings from breadwheat data
# into a object of class 'grouped_rankings'

names(breadwheat)

G <- to_rankings(breadwheat, 
                 items = c("variety_a","variety_b","variety_c"), 
                 rankings = c("overall_best","overall_worst"),
                 type = "tricot",
                 grouped.rankings = TRUE)


# compute temperature indices for the first 45 days in the plots

data("breadwheat_modis", package = "gosset")

temp <- temperature(breadwheat_modis,
                    day.one = breadwheat["planting_date"],
                    span = 45)


# combine grouped rankings with temperature indices
mydata <- cbind(G, temp)

# fit a pltree model using temperature data
mod <- pltree(G ~ ., data = mydata)


worst_regret(mod)

}
