% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpermute.R
\name{btpermute}
\alias{btpermute}
\title{Variable selection with pair comparisons}
\usage{
btpermute(
  contests = NULL,
  predictors = NULL,
  iterations = 15,
  seed = NULL,
  ...
)
}
\arguments{
\item{contests}{a data frame with pairwise binary contests with an 
id that should match with \code{predictors}}

\item{predictors}{a data frame with player predictors with an id 
that should match with \code{contests}}

\item{iterations}{integer, number of iterations to compute}

\item{seed}{integer, the seed for random number generation. If NULL (the default), 
gosset will set the seed randomly}

\item{...}{additional arguments passed to \pkg{BradleyTerry2} methods}
}
\description{
Permutation method to select variables for pair comparison data
}
\examples{
\donttest{  
data("kenyachoice", package = "gosset")

require("BradleyTerry2")

items <- unique(kenyachoice$contests$Item1)

mod <- btpermute(contests = kenyachoice$contests,
                 predictors = kenyachoice$predictors,
                 iterations = 15,
                 seed = 1)

mod
}
}
\seealso{
\code{\link{rank_binomial}}, \code{\link[BradleyTerry2]{BTm}}

Other model selection functions: 
\code{\link{forward}()}
}
\concept{model selection functions}
