% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{crossvalidation}
\alias{crossvalidation}
\title{Cross-validation}
\usage{
crossvalidation(
  formula,
  data,
  k = 10,
  folds = NULL,
  mean.method = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted,
of the form \eqn{ y ~ x1 + ... + xn }}

\item{data}{a data frame, list or environment 
(or object coercible by as.data.frame to a data frame) 
containing the variables in the model.}

\item{k}{an integer for the number of splits in the cross-validation}

\item{folds}{an optional vector specifying the folds in the cross-validation}

\item{mean.method}{a character for the method to calculate the mean of cross-validation estimators. 
Options are: 'equal', an arithmetic mean; 
'foldsize', a weighted mean by the size in each fold; 
'stouffer' a weighted Z-test. See references.}

\item{seed}{The seed for random number generation. If NULL (the default), gosset will set the seed randomly}

\item{...}{additional arguments passed to methods}
}
\value{
The cross-validation goodness-of-fit estimates, which are:
\item{AIC}{Akaike Information Criterion}
\item{deviance}{Model deviance}
\item{logLik}{Log-Likelihood}
\item{MaxLik}{Maximum likelihood pseudo R-squared}
\item{CraggUhler}{Cragg and Uhler's pseudo R-squared}
\item{Agresti}{Agresti pseudo R-squared}
}
\description{
Methods for measuring the performance of a predictive model on sets of test data.
}
\examples{
# Generalized Linear Models

data("airquality")

cv <- crossvalidation(Temp ~ Wind + Solar.R,
                      data = airquality,
                      k = 3,
                      seed = 999,
                      family = poisson())
                

\donttest{
# PlackettLuce Model
# beans data from PlackettLuce
library("PlackettLuce")
data("beans", package = "PlackettLuce")

G <- rank_tricot(data = beans,
                 items = c(1:3),
                 input = c(4:5),
                 additional.rank = beans[c(6:8)],
                 group = TRUE)

beans <- cbind(G, beans)

# take seasons as bins
k <- length(unique(beans$season))
folds <- as.integer(as.factor(beans$season))

cv <- crossvalidation(G ~ maxTN,
                      data = beans,
                      k = k,
                      folds = folds,
                      minsize = 100)

########################################
}

\dontrun{
# BradleyTerry Model
library("psychotree")
# Germany's Next Topmodel 2007 data from psychotree
data("Topmodel2007", package = "psychotree")

cv <- crossvalidation(preference ~ .,
                      data = Topmodel2007,
                      k = 5)
                
}
                
}
\references{
Agresti A. (2002). Categorical Data Analysis. John Wiley & Sons, Inc., Hoboken, NJ, USA. http://doi.wiley.com/10.1002/0471249688

Elder J. F. (2003). Journal of Computational and Graphical Statistics, 12(4), 853–864. https://doi.org/10.1198/1061860032733.

Whitlock M. C. (2005). Journal of Evolutionary Biology, 18(5), 1368–1373. https://doi.org/10.1111/j.1420-9101.2005.00917.x.
}
\seealso{
\code{\link[gnm]{gnm}}, \code{\link[PlackettLuce]{pltree}}, 
\code{\link[psychotree]{bttree}}
}
