% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{crossvalidation}
\alias{crossvalidation}
\title{Cross-validation}
\usage{
crossvalidation(
  formula,
  data,
  k = NULL,
  folds = NULL,
  mean.method = "Ztest",
  logLik.method = "Turner",
  seed = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be 
coerced to that class): a symbolic description of the model to be fitted,
of the form \eqn{y ~ x1 + ... + xn}}

\item{data}{a data frame (or object coercible by as.data.frame to a data frame)
containing the variables in the model}

\item{k}{an integer for the number of bins in the cross-validation}

\item{folds}{an optional vector or list of vectors specifying the \var{k}-folds in the cross-validation}

\item{mean.method}{a character for the method to calculate the mean of 
cross-validation estimators. 
Options are: 'equal', arithmetic mean; 
'foldsize', weighted mean by the size in each fold; 
'Ztest' weighted through Z-test. See references}

\item{logLik.method}{a character for the method to compute logLik 
options are "Hunter" or "Turner"}

\item{seed}{integer, the seed for random number generation. If NULL (the default), 
\pkg{gosset} will set the seed randomly}

\item{...}{additional arguments passed the methods of the chosen model}
}
\value{
an object of class \code{gosset_cv} with the cross-validation 
goodness-of-fit estimates, which are:
\item{AIC}{Akaike Information Criterion}
\item{deviance}{Model deviance}
\item{logLik}{Log-Likelihood}
\item{MaxLik}{Maximum likelihood pseudo R-squared}
\item{CraggUhler}{Cragg and Uhler's pseudo R-squared}
\item{Agresti}{Agresti pseudo R-squared}
\item{kendallTau}{the Kendall correlation coefficient, only for Plackett-Luce models}
}
\description{
Methods for measuring the performance of a predictive model on sets of 
test data in Bradley-Terry model from \pkg{psychotree}, Generalized Linear and 
Generalized Nonlinear models from \pkg{gnm}, and Plackett-Luce model from 
\pkg{PlackettLuce}
}
\examples{
# Generalized Linear Models
require("gnm")

data("airquality")

cv <- crossvalidation(Temp ~ Wind + Solar.R,
                      data = airquality,
                      k = 3,
                      seed = 999,
                      family = poisson())
                

\donttest{
# Folds as list 
require("caret")

fd <- createFolds(y = airquality$Day, k = 15, list = TRUE, returnTrain = TRUE)
 
crossvalidation(Temp ~ .,
                data = airquality,
                folds = fd,
                family = poisson(link = "log"))

########################################
  
# Plackett-Luce Model
require("PlackettLuce")

# beans data from PlackettLuce
data("beans", package = "PlackettLuce")

G <- rank_tricot(data = beans,
                 items = c(1:3),
                 input = c(4:5),
                 additional.rank = beans[c(6:8)],
                 group = TRUE)

beans <- cbind(G, beans)

# take seasons as bins
k <- length(unique(beans$season))
folds <- as.integer(as.factor(beans$season))

cv <- crossvalidation(G ~ maxTN,
                      data = beans,
                      k = k,
                      folds = folds,
                      minsize = 100)

########################################

# Bradley-Terry Model
require("psychotree")

# Germany's Next Topmodel 2007 data from psychotree
data("Topmodel2007", package = "psychotree")

cv <- crossvalidation(preference ~ .,
                      data = Topmodel2007,
                      k = 5)
                
}
                
}
\references{
Agresti A. (2002). Categorical Data Analysis. 
\url{http://doi.wiley.com/10.1002/0471249688}

Elder J. F. (2003). Journal of Computational and Graphical Statistics, 12(4), 853–864.
\url{https://doi.org/10.1198/1061860032733}

James G., et al. (2013). An Introduction to Statistical Learning: with Applications in R.
\url{https://doi.org/10.1007/978-1-4614-7138-7}

Whitlock M. C. (2005). Journal of Evolutionary Biology, 18(5), 1368–1373. 
\url{https://doi.org/10.1111/j.1420-9101.2005.00917.x}
}
\seealso{
\code{\link[psychotree]{bttree}}, 
\code{\link[gnm]{gnm}},
\code{\link[PlackettLuce]{pltree}}

Other model selection functions: 
\code{\link{bagging}()},
\code{\link{btpermute}()},
\code{\link{forward}()}
}
\author{
Kauê de Sousa, Jacob van Etten and David Brown
}
\concept{model selection functions}
