% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Forward stepwise regression for model selection}
\usage{
forward(
  formula,
  data,
  k = NULL,
  folds = NULL,
  select.by = NULL,
  akaike.weights = FALSE,
  ncores = NULL,
  packages = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model, if set as
\eqn{ y ~ . } all variables in \code{data} are used}

\item{data}{a data frame (or object coercible by as.data.frame to a data frame)
containing the variables in the model}

\item{k}{an integer for the number of bins in the cross-validation}

\item{folds}{an optional vector specifying the folds in the cross-validation}

\item{select.by}{a character for the goodness-of-fit statistical parameter to
select the models. Set as 'deviance' by default. Check 'Value' for details in outputs 
and goodness-of-fit options.}

\item{akaike.weights}{optional, logical object for averaging the goodness-of-fit 
coefficients with Akaike weights}

\item{ncores}{an integer for the number of cores to be used in the 
parallel computing. Set as 1 by default (no parallelisation)}

\item{packages}{character vector of packages that the parallel tasks depend on. If you use 
GLM or GNM you should call for \pkg{gnm}, for Plackett-Luce, use \pkg{PlackettLuce}, for 
Bradley-Terry, use \pkg{psychotree}}

\item{seed}{integer, the seed for random number generation. If NULL (the default), 
\pkg{gosset} will set the seed randomly}

\item{...}{additional arguments passed the methods of the chosen model}
}
\value{
an object of class \code{gosset_cv} with the cross-validation 
goodness-of-fit estimates, which are:
\item{AIC}{Akaike Information Criterion}
\item{deviance}{Model deviance}
\item{logLik}{Log-Likelihood}
\item{MaxLik}{Maximum likelihood pseudo R-squared}
\item{CraggUhler}{Cragg and Uhler's pseudo R-squared}
\item{Agresti}{Agresti pseudo R-squared}
Cross-validation estimates are computed using the fitted models on the validation samples.
}
\description{
Forward selection is a type of stepwise regression which begins with an empty
 model and adds in variables one by one. In each step one variable is added
 when it makes a single best improvement to the model.
}
\examples{
 

require("gnm")
require("foreach")
require("abind")
require("doParallel")

 
data("airquality")

mod <- forward(Temp ~ .,
               data = airquality,
               k = 3,
               select.by = "AIC",
               packages = "gnm",
               family = poisson(link = "log"))
               
\donttest{
##########################################
# Bradley-Terry Model
require("psychotree")

# Germany's Next Topmodel 2007 data from psychotree
data("Topmodel2007", package = "psychotree")

forward(preference ~ .,
        data = Topmodel2007,
        k = 5,
        packages = "psychotree",
        ncores = 5,
        alpha = 0.01,
        minsize = 30)
  
}
}
\seealso{
Other model selection functions: 
\code{\link{btpermute}()},
\code{\link{crossvalidation}()}
}
\concept{model selection functions}
