% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectsCM.R
\name{getProjectsCM}
\alias{getProjectsCM}
\title{Get ClimMob projects}
\usage{
getProjectsCM(key = NULL)
}
\arguments{
\item{key}{a character for the ClimMob user API key}
}
\value{
A data frame with the ClimMob projects 
\item{project_id}{the project unique id}
\item{name}{the project name}
\item{status}{the current status}
\item{creation_date}{the project's creation date}
}
\description{
Fetch the list of ClimMob projects using your API key
}
\examples{
\dontrun{ 
# This function will not work without your API key
library("gosset")
library("jsonlite")
library("httr")

my_key <- "my_api_key"

getProjectsCM(key = my_key)
}

}
\seealso{
\url{https://climmob.net/climmob3/}
}
