% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosset.R
\docType{package}
\name{gosset}
\alias{gosset}
\alias{gosset-package}
\title{Modelling Metadata and Experimental Citizen Science}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Toolkit for a workflow to analyse experimental citizen 
    science data, from collating data of different sources to model selection and 
    visualisation. The package is named after W.S. Gosset aka ‘Student’, a pioneer 
    of modern statistics in small sample experimental design and analysis.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{Development repository: 
 \url{https://github.com/agrdatasci/gosset}}
\item{Static documentation: 
 \url{https://agrdatasci.github.io/gosset/}}
\item{Report bugs: 
 \url{https://github.com/agrdatasci/gosset/issues}}
}
}
\author{
Kauê de Sousa and Jacob van Etten and Sam Dumble and 
 Nicolas Greliche and Jonathan Steinke and Svein Ø. Solberg
}
