% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosset.R
\docType{package}
\name{gosset}
\alias{gosset}
\alias{gosset-package}
\title{Tools for Data Analysis in Experimental Agriculture}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Toolkit to analyse experimental agriculture data, from data synthesis to model selection and visualisation. The package is named after W.S. Gosset aka ‘Student’, a pioneer of modern statistics in small sample experimental design and analysis.
}
\seealso{
\strong{Useful links:}
\itemize{
\item{gosset paper: 
 \doi{10.2139/ssrn.4236267}}
\item{Development repository: 
 \url{https://github.com/AgrDataSci/gosset}}
\item{Static documentation: 
 \url{https://AgrDataSci.github.io/gosset/}}
\item{Report bugs: 
 \url{https://github.com/AgrDataSci/gosset/issues}}
}
}
\author{
Kauê de Sousa and Jacob van Etten 
and David Brown and Jonathan Steinke
}
