% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/kendallTau.R
\name{kendallTau}
\alias{kendallTau}
\title{Kendall Tau correlation coefficient}
\usage{
kendallTau(x, y)
}
\arguments{
\item{x}{a matrix with rankings or model coefficients}

\item{y}{a matrix with rankings or model coefficients}
}
\value{
The Kendall correlation coefficient and the Effective N, which 
is the equivalent N needed if all items were compared to all items. 
Used for significance testing.
}
\description{
Compute Kendall Tau correlation coefficient between two matrices
}
\examples{

# check the correlation between observed rankings 
# and the predicted coefficients from PlackettLuce

library("PlackettLuce")

R <- matrix(c(1, 2, 4, 3,
              1, 4, 2, 3,
              1, 2, 4, 3,
              1, 2, 4, 3,
              1, 3, 4, 2,
              1, 4, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) <- LETTERS[1:4]

G <- grouped_rankings(as.rankings(R), 1:6)

mod <- pltree(G ~ 1, data = G)

preds <- predict(mod)

kendallTau(R, preds)

}
\seealso{
\code{\link[stats]{cor}}
}
