% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeinfo.R
\name{node_labels}
\alias{node_labels}
\alias{node_rules}
\alias{top_items}
\alias{plot.pltree}
\alias{plot_logworth}
\title{Get node labels and rules used in a party tree}
\usage{
node_labels(x)

node_rules(x)

top_items(x, top = 5)

\method{plot}{pltree}(x, log = TRUE, ref = NULL, ci.level = 0.95, ...)

plot_logworth(
  x,
  ci.level = 0.95,
  ref = NULL,
  multcomp = FALSE,
  levels = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class PlackettLuce}

\item{top}{an integer for the number of items to return}

\item{log}{logical, if \code{TRUE} log-worth coefficients are 
displayed instead of worth}

\item{ref}{optional, character for the reference item when
\var{log} = \code{TRUE}}

\item{ci.level}{the confidence interval level}

\item{...}{additional arguments passed to methods. See details}

\item{multcomp}{logical to add group letters}

\item{levels}{optional, a vector with the order of labels in axis Y}
}
\value{
a vector with the node labels, 
 a data.frame with node rules, a ggplot
}
\description{
Returns the covariates used to split a recursive
 partitioning tree and the rules that were applied 
 to build the tree
}
\details{
Argument multcomp = TRUE adds multi-comparison letters from multcompView
}
\examples{
\donttest{
library("PlackettLuce")
data("beans", package = "PlackettLuce")
G = rank_tricot(data = beans,
                 items = c(1:3),
                 input = c(4:5),
                 group = TRUE,
                 additional.rank = beans[c(6:8)])
 
pld = cbind(G, beans[,c("maxTN", "season", "lon")])

tree = pltree(G ~ maxTN + season + lon, data = pld)

node_labels(tree)

node_rules(tree)

top_items(tree)

plot(tree)

plot(tree, log = TRUE)
}
}
\author{
Kauê de Sousa
}
