% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{num2rank}
\alias{num2rank}
\title{Numbers into rankings}
\usage{
num2rank(object, ...)
}
\arguments{
\item{x}{vector with numbers of class "numeric" or "integer"}
}
\value{
a vector with ranked values where highest values are placed first
}
\description{
Group and convert numeric values into integer ranks
}
\examples{

x <- c(1, 2, 1.4, 3, 4.01)

is_wholenumber(x)

}
