% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nodes.R
\name{plot_nodes}
\alias{plot_nodes}
\title{Plot nodes from recursive partitioning trees}
\usage{
plot_nodes(object, labels = NULL, ...)
}
\arguments{
\item{object}{an object of class modelparty}

\item{labels}{optional, new labels for the axis X in the plot
by default}

\item{...}{additional arguments passed to ggplot2}
}
\value{
a list of plots with probabilities of winning and 
intervals based on quasi-standard errors
}
\description{
Make a ggplot2 chart from model-based recursive partitioning trees with quasi-variance
}
\examples{
\dontrun{

library("PlackettLuce")
library("qvcalc")
library("ggplot2")

data("breadwheat", package = "gosset")

# convert the tricot rankings from breadwheat data
# into a object of class 'grouped_rankings'
names(breadwheat)

G <- rank_PL(breadwheat, 
                 items = c("variety_a","variety_b","variety_c"), 
                 rankings = c("overall_best","overall_worst"),
                 type = "tricot",
                 grouped.rankings = TRUE)

# combine grouped rankings with lon lat from wheat dataset
mydata <- cbind(G, breadwheat[c("lon","lat")])

# fit a pltree model using lon and lat
mod <- pltree(G ~ ., data = mydata)

# plot nodes
plot_nodes(mod)

}

}
\seealso{
\code{\link[qvcalc]{qvcalc}}
}
