% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoR2.R
\name{pseudoR2}
\alias{pseudoR2}
\title{Pseudo R-squared}
\usage{
pseudoR2(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{additional arguments affecting the R-squared produced}
}
\value{
A data frame containing the pseudo R-squared coefficients:
\item{logLik}{log-likelihood}
\item{logLikNull}{Null log-likelihood}
\item{MaxLik}{maximum likelihood pseudo R-squared}
\item{CraggUhler}{Cragg and Uhler's pseudo R-squared}
\item{Agresti}{Agresti pseudo R-squared}
}
\description{
Regression coefficient to evaluate goodness-of-fit in a given model when 
ordinary least squares (OLS) are not available. The algorithm computes estimates 
from the maximum likelihood through an iterative process. These estimates are called 
'pseudo R-squared' because they look like 'R-squared' in the sense that they are on 
a similar scale (from 0 to 1), with higher values indicating better model fit.
}
\examples{

data("airquality")

mod <- glm(Temp ~ Wind + Solar.R,
           data = airquality,
           family = poisson())

pseudoR2(mod)

}
\references{
Agresti A. (2002). Categorical Data Analysis. John Wiley & Sons, Inc., Hoboken, NJ, USA. http://doi.wiley.com/10.1002/0471249688

Hunter D. R. (2004). The Annals of Statistics, 32(1), 384–406. http://www.jstor.org/stable/3448514
}
