% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoR2.R
\name{pseudoR2}
\alias{pseudoR2}
\title{Pseudo R-squared}
\usage{
pseudoR2(object, ...)
}
\arguments{
\item{object}{a model object for which pseudo R-squared is desired}

\item{...}{additional arguments affecting the R-squared produced}
}
\value{
A data frame containing the pseudo R-squared coefficients:
\item{logLik}{log-likelihood from the fitted model}
\item{logLikNull}{log-likelihood from the null model}
\item{MaxLik}{maximum likelihood pseudo R-squared}
\item{CraggUhler}{Cragg and Uhler's pseudo R-squared}
\item{Agresti}{Agresti pseudo R-squared}
}
\description{
Regression coefficient to evaluate goodness-of-fit in a given model
}
\examples{

data("airquality")

mod <- glm(Temp ~ Wind + Solar.R,
           data = airquality,
           family = poisson())

pseudoR2(mod)

#################################
\dontrun{

# Compute pseudo R-squared on a validation sample

n <- nrow(airquality)

s <- sample(1:n, n * 0.7)

train <- airquality[s,]
test <- airquality[-s,]

mod <- glm(Temp ~ Wind + Solar.R,
           data = train,
           family = poisson())

pseudoR2(mod, newdata = test)

}
}
