% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmental_indices.R
\name{rainfall}
\alias{rainfall}
\title{Rainfall indices}
\usage{
rainfall(object, day.one = NULL, span = NULL, index = NULL, ...)
}
\arguments{
\item{object}{a numeric vector of geographic coordinates (lonlat) or
a matrix containing the precipitation data.
When lonlat is used, the function makes a call to
\code{nasapower::get_power} to fetch and concatenate environmental data from NASA POWER
(\url{https://power.larc.nasa.gov/})}

\item{day.one}{a vector of class 'Date' for the starting date to capture the environmental data}

\item{span}{an integer or a vector with integers for the duration of the timespan to be captured}

\item{index}{optional, a character or a vector with characters for the indices to be computed. 
All indices are given by default}

\item{...}{additional arguments passed to methods}
}
\value{
A dataframe with selected indices. Options are:
\item{MLDS}{maximum length of consecutive dry days (r <  1 mm)}
\item{MLWS}{maximum length of consecutive wet days (r >= 1 mm)}
\item{R10mm}{number of heavy precipitation days (10 >= r < 20 mm)}
\item{R20mm}{number of very heavy precipitation days (r >= 20) }
\item{SDII}{simple daily intensity index (mean of wet days / total rainfall)}
\item{Rx1day}{maximum 1-day rainfall (mm)}
\item{Rx5day}{maximum 5-day rainfall (mm) }
\item{Rtotal}{total rainfall (mm) in wet days (R >= 1)}
}
\description{
Compute rainfall indices over a timespan
}
\examples{

# Compute indices using precipitation data from CHIRPS
data("breadwheat_chirps", package = "gosset")
data("breadwheat", package = "gosset")

# The rainfall indices for the first 50 days after sowing
rainfall(breadwheat_chirps, 
         day.one = breadwheat$planting_date, 
         span = 50)

########################################

# Add the first 7 days before sowing (residual precipitation)
rainfall(breadwheat_chirps, 
         day.one = breadwheat$planting_date, 
         span = 50,
         days.before = 7)

########################################

# Compute indices using NASA POWER

library("nasapower")
library("raster")

rainfall(breadwheat[c("lon","lat")], 
         day.one = breadwheat$planting_date, 
         span = 50)
         
}
