% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_binomial.R
\name{rank_binomial}
\alias{rank_binomial}
\title{Binary rankings}
\usage{
rank_binomial(object, drop.null = FALSE)
}
\arguments{
\item{object}{an object of class "rankings", "grouped_rankings" or "paircomp".}

\item{drop.null}{logical, an optional argument to remove players without comparisons}
}
\value{
a binary rank
}
\description{
Pairwise comparisons from a ranking object. Ties are not taken into account, then they are added as NA's.
}
\examples{
\\notrun{}
}
\seealso{
\code{\link[PlackettLuce]{rankings}}, \code{\link[psychotools]{paircomp}}
}
