% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_binomial.R
\name{rank_binomial}
\alias{rank_binomial}
\title{Binary rankings}
\usage{
rank_binomial(object, drop.null = FALSE)
}
\arguments{
\item{object}{an object of class "rankings", "grouped_rankings" or "paircomp".}

\item{drop.null}{logical, an optional argument to remove players without comparisons}
}
\value{
a binary rank
}
\description{
Binary comparisons from a ranking object. Ties are not taken into account, then they are added as NA's.
}
\examples{
# Rankings with 5 items randomly assigned

i <- as.data.frame(matrix(NA, nrow = 10, ncol = 5))
names(i) <- paste0("Item",1:5)

r <- as.data.frame(matrix(NA, nrow = 10, ncol = 5))
names(r) <- paste0("Position_Item",1:5)

for(s in 1:10) {
  i[s,] <- sample(LETTERS[1:5])
  r[s,] <- sample(1:5)
}

R <- rank_PL(items = i,
             input = r)

rank_binomial(R)

}
\seealso{
\code{\link[PlackettLuce]{rankings}}, \code{\link[psychotools]{paircomp}}
}
