% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_favourite.R
\name{summarise_favourite}
\alias{summarise_favourite}
\alias{summarise_favorite}
\alias{plot.gosset_fvrt}
\title{Favourability scores}
\usage{
summarise_favourite(object, ...)

\method{plot}{gosset_fvrt}(x, ...)
}
\arguments{
\item{object}{an object of class \code{rankings} or \code{grouped_rankings}}

\item{...}{additional arguments passed to methods. Not enabled yet}

\item{x}{an object of class \code{gosset_fvrt} for the plotting method. 
Generates a \code{ggplot} object that can be passed to any \pkg{ggplot2} method}
}
\value{
A data.frame with the descriptive statistics:
\item{N}{number of times the given item was evaluated}
\item{best}{relative number of times (in percentage) the given item was
 ranked as first}
\item{worst}{relative number of times (in percentage) the given item was 
ranked as worst}
\item{wins}{relative number of times (in percentage) the given item wins 
against the others}
\item{fav_score}{the favourability score, which is the difference between 
best and worst performance}
}
\description{
Summarise favourability scores from permutation contests
}
\examples{
data("breadwheat", package = "gosset")

R <- rank_tricot(data = breadwheat,
                 items = c("variety_a", "variety_b", "variety_c"),
                 input = c("overall_best", "overall_worst"))

fav <- summarise_favourite(R)

fav

}
\seealso{
Other summarise functions: 
\code{\link{summarise_agreement}()},
\code{\link{summarise_dominance}()},
\code{\link{summarise_victories}()}
}
\concept{summarise functions}
