% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{temperature}
\alias{temperature}
\title{Temperature indices}
\usage{
temperature(
  object,
  day.one = NULL,
  span = 150,
  timeseries = FALSE,
  intervals = 5,
  ...
)
}
\arguments{
\item{object}{a numeric vector of geographic coordinates (lonlat) or
an array with two dimensions containing the temperature data;
1st dimension contains the day temperature and 2nd dimension the night
temperature. When lonlat is used, the function makes a call to
\code{nasapower::get_power()} to fetch and concatenate environmental
data from NASA POWER (\url{https://power.larc.nasa.gov/})}

\item{day.one}{a vector of class \code{Date} for the starting date to 
capture the environmental data.}

\item{span}{an integer or a vector with integers for the duration 
of the timespan to be captured}

\item{timeseries}{logical, \code{FALSE} for a single point time series
observation or \code{TRUE} for a time series based on \var{intervals}}

\item{intervals}{integer no lower than 5, for the days intervals when
\var{timeseries} = \code{TRUE}}

\item{...}{additional arguments passed to methods}
}
\value{
A dataframe with temperature indices:
\item{maxDT}{maximun day temperature (degree Celsius)}
\item{minDT}{minimum day temperature (degree Celsius)}
\item{maxNT}{maximun night temperature (degree Celsius)}
\item{minNT}{minimum night temperature (degree Celsius) }
\item{DTR}{diurnal temperature range (mean difference between DT 
and NT (degree Celsius)) }
\item{SU}{summer days, number of days with maximum temperature > 
30 (degree Celsius)}
\item{TR}{tropical nights, number of nights with maximum 
temperature > 25 (degree Celsius) }
}
\description{
Compute temperature indices over a timespan
}
\examples{
\donttest{

library("gosset")
library("nasapower")

# random geographic locations around bbox(11, 12, 55, 58)
lonlat <- data.frame(lon = runif(10, 11, 12),
                     lat = runif(10, 55, 58))

# random planting dates around 2018-05-15 and 2018-05-20
pdates <- as.integer(runif(10, 17666, 17670))
pdates <- as.Date(pdates, origin = "1970-01-01")

# get temperature indices for the first 40 days
temperature(lonlat,
            day.one = pdates,
            span = 40)
}
}
\references{
Aguilar E., et al. (2005). Journal of Geophysical Research, 
110(D23), D23107. \cr\url{https://doi.org/10.1029/2005JD006119}

Kehel Z., et al. (2016). Identifying Climate Patterns during 
the Crop-Growing Cycle from 30 Years of CIMMYT Elite Spring 
Wheat International Yield Trials. In: Applied Mathematics and 
Omics to Assess Crop Genetic Resources for Climate Change Adaptive 
Traits (eds Bari A., Damania A. B., Mackay M., Dayanandan S.), 
pp. 151–174. CRC Press.

Sparks A. H. (2018). Journal of Open Source Software, 3(30), 1035. 
\cr\url{https://doi.org/10.21105/joss.01035}
}
\seealso{
Other climatology functions: 
\code{\link{ETo}()},
\code{\link{GDD}()},
\code{\link{rainfall}()}
}
\concept{climatology functions}
