% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmental_indices.R
\name{temperature}
\alias{temperature}
\title{Temperature indices}
\usage{
temperature(object, day.one = NULL, span = NULL, index = NULL, ...)
}
\arguments{
\item{object}{a numeric vector of geographic coordinates (lonlat) or
an array with two dimensions containing the temperature data;
1st dimension contains the day temperature and 2nd dimension the night temperature.
When lonlat is used, the function makes a call to
\code{nasapower::get_power} to fetch and concatenate environmental data from NASA POWER
(\url{https://power.larc.nasa.gov/})}

\item{day.one}{a vector of class 'Date' for the starting date to capture the environmental data}

\item{span}{an integer or a vector with integers for the duration of the timespan to be captured}

\item{index}{optional, a character or a vector with characters for the indices to be computed. 
All indices are given by default}

\item{...}{additional arguments passed to methods}
}
\value{
A dataframe with selected indices. Options are:
\item{maxDT}{maximun day temperature (degree Celsius)}
\item{minDT}{minimum day temperature (degree Celsius)}
\item{maxNT}{maximun night temperature (degree Celsius)}
\item{minNT}{minimum night temperature (degree Celsius) }
\item{DTR}{diurnal temperature range (mean difference between DT and NT (degree Celsius)) }
\item{SU}{summer days, number of days with maximum temperature > 30 (degree Celsius)}
\item{TR}{tropical nights, number of nights with maximum temperature > 25 (degree Celsius) }
}
\description{
Compute temperature indices over a timespan
}
\examples{
# Compute indices using temperature data from MODIS MYD11A1
# (http://dx.doi.org/10.5067/MODIS/MYD11A2.006)
# each row in breadwheat corresponds to the same rows in breadwheat_temp,
# first layer is the day temperature and second layer is the night temperature

data("breadwheat_temp", package = "gosset")
data("breadwheat", package = "gosset")

# compute all temperature indices for the first 30 days in the plots
temperature(breadwheat_temp,
            day.one = breadwheat["planting_date"],
            span = 30)


########################################

# Temperature indices from NASA POWER

library("nasapower")
library("raster")

temperature(breadwheat[c("lon","lat")], 
            day.one = breadwheat["planting_date"], 
            span = 40)

}
