% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_PL.R
\name{to_rankings}
\alias{to_rankings}
\title{Plackett-Luce rankings}
\usage{
to_rankings(...)
}
\arguments{
\item{...}{additional arguments passed to methods. See Details}
}
\description{
Create an object of class "rankings" from a dataframe or matrix
}
